+++ fx-application/src/main/java/com/dua3/fx/application/FxApplication.java
@@ -18,6 +18,7 @@ import com.dua3.fx.util.Dialogs;
 import com.dua3.fx.util.controls.AboutDialog;
 import com.dua3.utility.data.Pair;
 import com.dua3.utility.lang.LangUtil;
+import com.dua3.utility.lang.Platform;
 import com.dua3.utility.text.TextUtil;
 import javafx.application.Application;
 import javafx.beans.value.ChangeListener;
@@ -47,53 +48,108 @@ import java.util.regex.Pattern;
 
 public abstract class FxApplication<A extends FxApplication<A, C>, C extends FxController<A, C>> extends Application {
 
-    // keep a strong reference to the root logger if needed, so that the configuration won't get lost
-    private static Logger rootLogger = null;
-    
-    private static synchronized Logger getRootLogger() {
-        if (rootLogger==null) {
-            rootLogger = Logger.getLogger("");
-        }
-        return rootLogger;
-    }
-    
-    public static
-    <A extends FxApplication<A, C>, C extends FxController<A, C>>
-    void launchApplication(Class<A> cls, String... args) {
-        String packageName = cls.getPackageName();
-
-        String logFile = System.getenv(packageName+".log_file");
-        String logLevel = System.getenv(packageName+".log_level");
-        if (logFile!=null) {
-            Level lvl = logLevel==null ? Level.INFO : Level.parse(logLevel);
-            try {
-                FileHandler fh = new FileHandler(logFile, false);
-                fh.setLevel(lvl);
-                fh.setFormatter(new SimpleFormatter());
-                getRootLogger().addHandler(fh);
-                getRootLogger().setLevel(lvl);
-            } catch (IOException e) {
-                Logger.getLogger(FxApplication.class.getName()).log(Level.WARNING, "cannot log to file", e);
-            }
-        }
-
-        Logger.getLogger(FxApplication.class.getName()).fine(() -> {
-            try (Formatter msg = new Formatter()) {
-                msg.format("Program arguments: %n");
-                for (int i = 0; i < args.length; i++) {
-                    msg.format("%n  %2d: \"%s\"", i, args[i]);
-                }
-                return msg.toString();
-            }
-        });
-        
-        Application.launch(cls, args);
-    }
-    
     /**
      * Logger
      */
     private static final Logger LOG = Logger.getLogger(FxApplication.class.getName());
+
+    /**
+     * Start application.
+     * This method is a drop-in replacement for `Application.launch(cls, args)`.
+     * <ul>
+     *     <li><strong>Comand line arguments</strong> are re-parsed on windows, for details see 
+     *     {@link #reparseCommandLine(String[])}.
+     * </ul>
+     * 
+     * @param cls
+     *  the application class
+     * @param args
+     *  the comand line arguments
+     * @param <A>
+     *  the application class
+     * @param <C>
+     *  the controller class
+     */
+    public static
+    <A extends FxApplication<A, C>, C extends FxController<A, C>>
+    void launchApplication(Class<A> cls, String... args) {
+        LOG.fine(() -> "arguments: "+Arrays.toString(args));
+        
+        var _args = reparseCommandLine(args);
+        if (_args != args) {
+            LOG.info(() -> "arguments have been re-parsed!");
+            LOG.fine(() -> "arguments: "+Arrays.toString(_args));
+        }
+        
+        Application.launch(cls, reparseCommandLine(_args));
+    }
+
+    /**
+     * Re-parse command line arguments.
+     * <ul>
+     * <li><strong>Windows:</strong>
+     * At least when using a jpackaged application with file-associations, command line arguments get messed up
+     * when the application start is the result of double-clicking on a registered file type.
+     * The command line args are split on whitespace, i. e. paths containing spaces
+     * will be split into multiple parts. This method tries to restore what was probably meant.
+     * It works by iterating over the given array of argumnents like this:
+     * <pre>
+     * let arg = "" 
+     * for each s in args:
+     *
+     *   if s starts with "--" // start of an option
+     *   or s starts with "[letter]:\" or "[letter]:/" // probable file path 
+     *   then 
+     *     append arg to the output array
+     *     arg = ""
+     *     
+     *   if arg != ""
+     *     arg = arg + " "
+     *     
+     *   arg = arg + s
+     * push arg to the output array
+     * </pre>
+     * <li><strong>Other platforms:</strong>
+     * The input array is returned without any changes.
+     * </ul>
+     * @param args
+     *   the command line arguments
+     * @return
+     *   the re-parsed argument array
+     * @deprecated this method is a workaround that will be removed once the underlying issue is fixed in the JDK.
+     */
+    @Deprecated
+    public static String[] reparseCommandLine(String[] args) {
+        if (!Platform.isWindows()) {
+            return args;
+        }
+
+        if (args.length<2) {
+            return args;
+        }
+
+        List<String> argL = new ArrayList<>();
+        StringBuilder arg = new StringBuilder();
+        for (String s:args) {
+            // split if s contains spaces, starts with a double dash, or a windows path
+            if (s.indexOf(' ')>=0 || s.matches("^(--|[a-zA-Z]:[/\\\\]).*")) {
+                if (arg.length()>0) {
+                    argL.add(arg.toString());
+                }
+                arg.setLength(0);
+            }
+
+            if (arg.length()>0) {
+                arg.append(' ');
+            }
+            arg.append(s);
+        }
+        if (arg.length()>0) {
+            argL.add(arg.toString());
+        }
+
+        return argL.toArray(String[]::new);
+    }
 
     // - constants -